requirements.txt
streamlit
tensorflow
opencv-python-headless
faiss-cpu
scikit-image
Pillow
